Move activities to seperate package.
