+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".TokenActivity"
+            android:name=".activities.TokenActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/SplashTheme"
@@ -38,7 +38,7 @@
         </activity>
 
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:label="@string/app_name"
             android:screenOrientation="portrait"
             android:theme="@style/AppTheme.NoActionBar"
@@ -55,7 +55,7 @@
         </activity>
 
         <activity
-            android:name=".CourseDetailActivity"
+            android:name=".activities.CourseDetailActivity"
             android:screenOrientation="portrait" />
 
         <service
@@ -73,7 +73,7 @@
                 android:resource="@xml/provider_paths" />
         </provider>
         <activity
-            android:name=".InfoActivity"
+            android:name=".activities.InfoActivity"
             android:screenOrientation="portrait" />
 
     </application>
