+++ app/src/main/java/crux/bphc/cms/fragments/SearchCourseFragment.java
@@ -21,7 +21,7 @@ import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import java.util.ArrayList;
 import java.util.List;
 
-import crux.bphc.cms.CourseDetailActivity;
+import crux.bphc.cms.activities.CourseDetailActivity;
 import crux.bphc.cms.R;
 import helper.ClickListener;
 import helper.MoodleServices;
