+++ app/src/main/res/layout/activity_token.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="crux.bphc.cms.TokenActivity">
+    tools:context="crux.bphc.cms.activities.TokenActivity">
 
     <ImageView
         android:id="@+id/background"
